deleted some unnecesary code, separated prompt clas to a file
