+++ PokemonGo.RocketAPI.Console/PokemonGo.RocketAPI.Console.csproj
@@ -156,6 +156,7 @@
       <DependentUpon>Pokemons.cs</DependentUpon>
     </Compile>
     <Compile Include="Program.cs" />
+    <Compile Include="Prompt.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Properties\Resources.Designer.cs">
       <AutoGen>True</AutoGen>
